Move render_default to OpenGLRenderer
